+++ database/migrations/2020_02_07_064509_create_users_last_data_table.php
@@ -17,7 +17,7 @@ class CreateUsersLastDataTable extends Migration
             $table->id();
             $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('role_id')->unsigned();
+            $table->bigInteger('role_id')->unsigned();
             $table->foreign('role_id')->references('id')->on('roles');
             $table->string('name');
             $table->string('email');
